Closes #28: rearrange modules
Move parsers and their utilities and tests to parsers/ submodule
