+++ src/grblogtools/parsers/single_log.py
@@ -1,9 +1,9 @@
-from grblogtools.continuous_parser import ContinuousParser
-from grblogtools.header_parser import HeaderParser
-from grblogtools.nodelog import NodeLogParser
-from grblogtools.norel import NoRelParser
-from grblogtools.presolve_parser import PresolveParser
-from grblogtools.termination import TerminationParser
+from grblogtools.parsers.continuous import ContinuousParser
+from grblogtools.parsers.header import HeaderParser
+from grblogtools.parsers.nodelog import NodeLogParser
+from grblogtools.parsers.norel import NoRelParser
+from grblogtools.parsers.presolve import PresolveParser
+from grblogtools.parsers.termination import TerminationParser
 
 
 class SingleLogParser:
